--- src/functionalTest/java/uk/gov/hmcts/reform/wataskconfigurationapi/CreatorObjectMapper.java
@@ -1,25 +1,0 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-
-public class CreatorObjectMapper {
-    private CreatorObjectMapper() {
-    }
-
-    public static String asJsonString(final Object obj) {
-        return jsonString(obj, new ObjectMapper().setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE));
-    }
-
-    public static String asCamundaJsonString(final Object obj) {
-        return jsonString(obj, new ObjectMapper());
-    }
-
-    private static String jsonString(Object obj, ObjectMapper mapper) {
-        try {
-            return mapper.writeValueAsString(obj);
-        } catch (Exception e) {
-            throw new IllegalArgumentException(e);
-        }
-    }
-}
